Fix <action :delete> for mysql_config.
I tripped over that issue the other day. I had created something like:
<if 'server-id' in node['mysql']['config'] create a config file, else delete it>.
Creation works nicely, but deleting the config didn't work, which was especially
fun because the mysqld refused to start with that config, so everything went
belly up :)

Figures, the problem is that the template in the :create-action appends the
extension ".cnf" to the configuration name, but the :delete-action appends the
extension ".conf" to the configuration name.

As far as I can see, there are no tests for the delete action, so I don't think
I broke tests.
Style checks still run properly..