Mocked service_resource_providers in ubuntu specs.
I've just put in another pull request and figured I might as well take a look at the
failing tests so it's easier to change this cookbook.

Overall, when printing a lot of stuff in the provider detection, I noticed that the
detected service_resource_providers were all over the place for the ubuntu tests.
When I cross-checked with the other tests, they all mock the service_resource_providers
to be the expected output on the target platform. That makes sense, because the unit
tests should be independent from the workstation they run on.

Thus, I've added mocks such that service_resource_providers in the ubuntu tests returns
[:ubuntu, :upstart], which triggers the upstart providers and fixes a bunch of tests.