Initial support for OpenSUSE 13
Sure is Systemd now, but it uses a different path for the unit files.  Add a helper to determine the path based on the platform. Also use the major version of suse so we can support all the 13s and avoid updating this cookbook every time a new release comes out.