Add dependency on apparmor cookbook
This fixes #660.

We test for `node['apparmor']['disable']` in `libraries/mysql_service_base.rb`
and fails on systems which doesn't include the apparmor cookbook. If we're
depending on this attribute, we should also be depending on the cookbook which
uses it, so this adds that dependency.

In addition, update the comment in the test for why we disable apparmor for our
integration tests.

Signed-off-by: Lance Albertson <lance@osuosl.org>